Adiciona implementação do CorujaTrainer e suporte a treinamento K-Fold com early stopping, além de modularização do código e adição de transformações de imagem.
